Music_Generator/requirements.txt
requests==2.26.0
azure-identity==1.13.0
azure-keyvault-secrets==4.7.0